backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
Pillow==9.0.0
python-dotenv==1.0.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-cleanup==8.1.0
django-filter==2.4.0
gunicorn==21.2.0
uharfbuzz==0.39.1
drf_extra_fields==3.7.0
reportlab==4.3.1
flake8
isort